renaming OGB to OGBG like the dataset. fixing and documenting we return the global batch size. returning per-example losses in OGBG.
