+++ baselines/imagenet/imagenet_jax/submission.py
@@ -13,6 +13,7 @@ from algorithmic_efficiency import spec
 
 
 def get_batch_size(workload_name):
+  # Return the global batch size.
   del workload_name
   return 128
