+++ baselines/imagenet/imagenet_pytorch/submission.py
@@ -10,6 +10,7 @@ from algorithmic_efficiency import spec
 
 
 def get_batch_size(workload_name):
+  # Return the global batch size.
   batch_sizes = {'imagenet_pytorch': 128}
   return batch_sizes[workload_name]
