+++ baselines/librispeech/librispeech_pytorch/submission.py
@@ -10,6 +10,7 @@ ctc_loss = torch.nn.CTCLoss(blank=0, reduction="none")
 
 
 def get_batch_size(workload_name):
+  # Return the global batch size.
   batch_sizes = {"librispeech_pytorch": 8}
   return batch_sizes[workload_name]
