+++ baselines/mnist/mnist_jax/submission.py
@@ -12,6 +12,7 @@ from algorithmic_efficiency import spec
 
 
 def get_batch_size(workload_name):
+  # Return the global batch size.
   batch_sizes = {'mnist_jax': 1024}
   return batch_sizes[workload_name]
